Create a tensor TensorExpression
