+++ src/DataStructures/Tensor/Expressions/CMakeLists.txt
@@ -10,5 +10,6 @@ spectre_target_headers(
   Evaluate.hpp
   LhsTensorSymmAndIndices.hpp
   Product.hpp
+  TensorAsExpression.hpp
   TensorExpression.hpp
   )
